+++ Source/Modules/Catalog/Gapotchenko.FX/Threading/ExecuteOnce.cs
@@ -5,6 +5,7 @@
 // Year of introduction: 2019
 
 using System.Diagnostics;
+using System.Runtime.CompilerServices;
 using System.Security.Permissions;
 
 namespace Gapotchenko.FX.Threading;
@@ -45,9 +46,30 @@ public struct ExecuteOnce
     }
 
     /// <summary>
+    /// Initializes a new instance of the <see cref="ExecuteOnce"/> struct.
+    /// </summary>
+    /// <param name="action">The action.</param>
+    /// <param name="syncLock">
+    /// A <see cref="Lock"/> object used as the mutually exclusive lock for action execution.
+    /// When the given value is <see langword="null"/>, an unique synchronization lock object is used.
+    /// </param>
+    public ExecuteOnce(Action action, Lock? syncLock) :
+#pragma warning disable CS9216 // A value of type 'System.Threading.Lock' converted to a different type will use likely unintended monitor-based locking in 'lock' statement.
+        this(action, (object?)syncLock)
+#pragma warning restore CS9216 // A value of type 'System.Threading.Lock' converted to a different type will use likely unintended monitor-based locking in 'lock' statement.
+    {
+    }
+
+    /// <summary>
     /// Ensures that the action was executed.
     /// </summary>
-    public void EnsureExecuted() => LazyInitializerEx.EnsureInitialized(ref m_SyncLock, ref m_Action);
+    public void EnsureExecuted()
+    {
+        if (m_SyncLock is Lock)
+            LazyInitializerEx.EnsureInitialized(ref Unsafe.As<object?, Lock?>(ref m_SyncLock), ref m_Action);
+        else
+            LazyInitializerEx.EnsureInitialized(ref m_SyncLock, ref m_Action);
+    }
 
     /// <summary>
     /// Gets a value indicating whether the action was executed.